requirements.txt
Django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
python-decouple==3.5
drf-spectacular==0.21.1
drf-spectacular-sidecar==2021.12.13
celery[redis]==5.2.3
psycopg2-binary==2.9.3
django-celery-results==2.3.0
django-celery-beat
flower==1.0.0
gunicorn==20.1.0
sentry-sdk==1.5.8
